+++ CHANGES.rst
@@ -5,7 +5,13 @@
 5.4.1 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Allow the filename for the logging configuration to be specified
+  via the envvar ``ZOPE_TESTRUNNER_LOG_INI``.
+  If not defined, the configuration continues to be locked for
+  in file ``log.ini`` of the current working directory.
+  Remember the logging configuration file in envvar
+  ``ZOPE_TESTRUNNER_LOG_INI`` to allow spawned child processes
+  to recreate the logging configuration.
 
 
 5.4.0 (2021-11-19)